Fix some embarrassing bugs, add the --exec option and now Tapir can run Plumage's test suite!
Various bugs in parsing unrecognized lines in TAP were fixed, and now we
are able to run the Plumage test suite with Tapir.

It appears that for the current Plumage test suite of 64 tests in 4
files, Tapir is the same speed as nqpTAP (which was somewhat unexpected),
while having more features, fewer bugs and lots of tests. t/harness in
the Plumage repo has no tests.