Add some tests for parsing out of order tests, mostly passing.
Currently we do not count any tests that are out of order as passed,
so test files with out of order tests do not pass, but this is currently
dumb luck. The parse_tapstream method needs to be modified to detect
out of order tests and still count which pass and fail.