Found the Bug in our Hunt
There was a bug where the final tally of failed tests was incorrect,
because only 1 failing test was counted per file, instead of the number
of tests that actually failed in that file. There are no tests to change
because the harness code that prints results is not tested. This sucks,
and if you don't like it, patches welcome!